pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "softcat-ai"
version = "0.1.0"
description = "A conversational agent spawner. You describe it. The cat builds it."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Valori" },
]
keywords = ["ai", "agents", "claude", "automation", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.42.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "httpx>=0.27.0",
    "python-crontab>=3.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
softcat = "softcat.cli:cli"

[project.urls]
Homepage = "https://softcat.ai"
Repository = "https://github.com/valori-collective/softcat"
Documentation = "https://softcat.ai/lab/softcat-cli"

[tool.hatch.build.targets.wheel]
packages = ["softcat"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
